+++ Engine/src/graphics/buffer/GPUBuffer.cpp
@@ -3,7 +3,7 @@
 namespace Ghurund {
     void GPUBuffer::set(Graphics& graphics, CommandList& commandList, unsigned int bindSlot) {
         ID3D12Resource* constantBufferUploadHeap = nullptr; // this is the memory on the gpu where our constant buffer will be placed.
-        Pointer* resourcePointer = graphics.ResourceFactory.create(D3D12_HEAP_TYPE_UPLOAD, CD3DX12_RESOURCE_DESC::Buffer(align<unsigned int>(buffer->Size, 1024 * 64)), D3D12_RESOURCE_STATE_GENERIC_READ, &constantBufferUploadHeap);
+        Pointer* resourcePointer = graphics.ResourceFactory.create(D3D12_HEAP_TYPE_UPLOAD, CD3DX12_RESOURCE_DESC::Buffer(align<size_t>(buffer->Size, 1024 * 64)), D3D12_RESOURCE_STATE_GENERIC_READ, &constantBufferUploadHeap);
 
 #ifdef _DEBUG
         if (resourcePointer == nullptr) {
@@ -14,7 +14,7 @@ namespace Ghurund {
 
         D3D12_CONSTANT_BUFFER_VIEW_DESC cbvDesc = {};
         cbvDesc.BufferLocation = constantBufferUploadHeap->GetGPUVirtualAddress();
-        cbvDesc.SizeInBytes = align<unsigned int>(buffer->Size, 256);    // CB size is required to be 256-byte aligned.
+        cbvDesc.SizeInBytes = align<UINT>((UINT)buffer->Size, 256);    // CB size is required to be 256-byte aligned.
         graphics.Device->CreateConstantBufferView(&cbvDesc, descHandle.getCpuHandle());
 
         CD3DX12_RANGE readRange(0, 0);
